+++ ChangeLog.md
@@ -2,8 +2,7 @@
 
 ## Version [0.5.0](https://github.com/MC2IT/Agicap.net/compare/v0.4.0...v0.5.0)
 - Added the `NestedList` class.
-- Added the `Payments.SynchronizedBeneficiary` class.
-- Added the `BeneficiarySynchronizationErrorCode` enumeration.
+- **Payments:** added the beneficiary synchronization API.
 
 ## Version [0.4.0](https://github.com/MC2IT/Agicap.net/compare/v0.3.0...v0.4.0)
 - Added the `HttpResponseException` class.