+++ ChangeLog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Version [0.6.0](https://github.com/MC2IT/Agicap.net/compare/v0.5.0...v0.6.0)
+- Added the `SynchronizedBeneficiary.BankAccount` property.
+
 ## Version [0.5.0](https://github.com/MC2IT/Agicap.net/compare/v0.4.0...v0.5.0)
 - Added the `NestedList` class.
 - **Payments:** added the beneficiary synchronization API.